All Other Necessary Files
Application Manager (.h and .cpp):

The Most Important .cpp Files. An Object from the Application Manager is made in the 'main.cpp' and function 'ApplicationManager::Run' is called, the Program Loop is initiated there.

////////////////////////////////////////////////////////

DrawingAreaAction (.h and .cpp):
Should've been placed inside the Actions Folder for Arrangement purposes, but was placed here during working on the Project.
PT-project.sln:
Run this on Visual Studio to get to Use the Program.

////////////////////////////////////////////////////////

Other Auxillary Files are Also Included.